requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
langgraph
pydantic
loguru
tiktoken
langchain-mcp-adapters
